Add files via upload
This is a simple hangman game, where you get to guess each letter of a word and avoid the diagram of a hanged man from being completed, as each wrong guess leads to the diagram getting completed, and its completion leads to the player losing the game. Now let's see what you got will you be able to gues the word or will you let a man get hanged.